requirement.txt
absl-py==1.1.0
aim==3.11.0
aim-ui==3.11.0
aimrecords==0.0.7
aimrocks==0.2.1
aiofiles==0.8.0
aiohttp==3.8.1
aiosignal==1.2.0
alembic==1.8.0
anyio==3.6.1
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
async-timeout==4.0.2
attrs==21.4.0
autograd==1.4
autoray==0.3.2
Babel==2.10.3
backcall==0.2.0
base58==2.0.1
beautifulsoup4==4.11.1
black==22.6.0
bleach==5.0.1
brotlipy==0.7.0
cachetools==5.2.0
certifi==2022.6.15
click==8.1.3
cmake==3.22.5
cutensor==1.5.0.3
cycler==0.11.0
Cython==0.29.30
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
docker-pycreds==0.4.0
entrypoints==0.4
estraces==1.7.1
executing==0.8.3
fastapi==0.67.0
fastjsonschema==2.15.3
fastrlock==0.8
filelock==3.7.1
flaky==3.7.0
fonttools==4.33.3
frozenlist==1.3.0
fsspec==2022.5.0
future==0.18.2
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.8.0
google-auth-oauthlib==0.4.6
greenlet==1.1.2
grpcio==1.42.0
h11==0.13.0
h5py==2.10.0
idna @ file:///tmp/build/80754af9/idna_1637925883363/work
importlib-metadata==4.12.0
importlib-resources==5.8.0
iniconfig==1.1.1
ipykernel==6.15.0
ipython==8.4.0
ipython-genutils==0.2.0
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
json5==0.9.8
jsonschema==4.6.1
jupyter-client==7.3.4
jupyter-core==4.10.0
jupyter-server==1.18.0
jupyterlab==3.4.3
jupyterlab-pygments==0.2.2
jupyterlab-server==2.14.0
kiwisolver==1.4.3
llvmlite==0.38.1
Mako==1.2.0
Markdown==3.3.7
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mistune==0.8.4
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186064646/work
mkl-service==2.4.0
multidict==6.0.2
mypy-extensions==0.4.3
nbclassic==0.4.0
nbclient==0.6.6
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
networkx==2.8.4
ninja==1.10.2.3
notebook-shim==0.1.0
numba==0.55.2
numpy @ file:///opt/conda/conda-bld/numpy_and_numpy_base_1652801679809/work
oauthlib==3.2.0
packaging==21.3
pandas==1.4.3
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.9.0
pathtools==0.1.2
PennyLane==0.24.0
PennyLane-Lightning==0.24.0
PennyLane-Lightning-GPU==0.24.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==2.5.2
pluggy==1.0.0
prometheus-client==0.14.1
promise==2.3
prompt-toolkit==3.0.29
protobuf==3.19.4
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
py3nvml==0.2.7
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic==1.9.1
pyDeprecate==0.3.2
Pygments==2.12.0
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
pyparsing==3.0.9
pyrsistent==0.18.1
PySocks @ file:///tmp/build/80754af9/pysocks_1605305779399/work
pytest==7.1.2
pytest-mock==3.8.1
python-dateutil==2.8.2
pytorch-lightning==1.6.4
pytz==2022.1
PyYAML==6.0
pyzmq==23.2.0
requests @ file:///opt/conda/conda-bld/requests_1641824580448/work
requests-oauthlib==1.3.1
RestrictedPython==5.2
retworkx==0.11.0
rsa==4.8
scared==0.8.13
scikit-learn==1.1.1
scipy==1.8.1
semantic-version==2.6.0
Send2Trash==1.8.0
sentry-sdk==1.6.0
setproctitle==1.2.3
shortuuid==1.0.9
smmap==5.0.0
sniffio==1.2.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.39
stack-data==0.3.0
starlette==0.14.2
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
terminado==0.15.0
threadpoolctl==3.1.0
tinycss2==1.1.1
toml==0.10.2
tomli==2.0.1
torch==1.11.0
torch-tb-profiler==0.4.0
torchaudio==0.11.0
torchmetrics==0.9.1
torchvision==0.12.0
tornado==6.1
tqdm==4.64.0
traitlets==5.3.0
trsfile==0.3.2
uvicorn==0.18.1
wandb==0.12.20
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.3
Werkzeug==2.1.2
xmltodict==0.13.0
yarl==1.7.2
zipp==3.8.0
